MDL-61292 tool_policy: Improve site policies checks in require_login()
The patch adds support for a new pre-config flag NO_SITEPOLICY_CHECK.
This constant should be defined and set to true if we should not check
the user's policyagreed status during the require_login().

This is for pages where the user actually accepts the site policies and
helps to avoid the redirect loop.

Additionally, we no longer check for policies in AJAX requests as those
may happen on policy agreement pages, too.