Implement the Moodle core version checking
The patch adds a warning to be displayed at the administration UI to
warn the admins that a special attention should ba paid to the installed
additional plugin.

If the admin forgets to remove the legacy files during the upgrade, the
environment check should fail and block the upgrade with another
explanation.